config handling completed
